fix checking of required options
Slop was checking if the list of arguments contained the long-name
version of an option to check if a required option was indeed supplied.

This obviously doesn't work if a required option was given by using
it's short name instead.

A saner way of checking if an option was given is by checking if the
argument_value of the option is not nil, as there is no way for the
user to supply nil as a valid value.