New support_dash option. Will treat --database-name like --database_name
In your ruby code we really prefer underscores. This way we can write:

    on :database_name=

Instead of the uglier:

    on 'database-name='

Also it means we can use the magic present? method. So we can do:

    opts.database_name?

Instead of the uglier:

    opts.present? 'database-name'

But using an underscore to make our Ruby prettier makes our command line
options uglier. We end up with a mix of - and _ which is confusing to users.

This patch adds a new option called :support_dash which will allow users to
type --database-name but still allow the ruby code to specify :database_name
and opts.database_name? It works similar to the existing case option that allows
options to be case insensitive.