ensure range 'typecast' allows negative range values
When using `:optional_argument => true` with a negative range
(-1..10 for example) #process_item assumes that the option argument
looks like an option, thus setting the argument to nil instead of
the negative range.

This change now allows anything prefixed with a `-` following by a digit
could be a viable option argument.

In the future we should probably allow "numeric options" (of course
this commit would half rule that out, when using optional arguments,
at least). We could enable and disable this via configuration options
sent to Slop.new

closes #65