ensure multiple switches respect count increases
Before this change, a multiple switch would for a value of 1 to
the option. That is, -aaa would force the option 'a' to have the
count of 1 and the argument 'aa'. But what if you don't want the
argument and you're expecting the count? This should return 3.
This commit fixes this issue by instead of setting the value to
1, it'll increase it by 1.

We also need to ensure count is **always** set before we do so, I
had the thought I was setting 1 explicitly for a reason.

ref #65