avoid using #present? when checking for an options present in autocreate
the present? method checks that an option exists AND has been used in the
argument list. autocreate works should just check for an options existence
not if the option has been used or not, if the option already exists,
autocreate should skip it. Before this commit, that was not the case

closes #77 again