Allow subclass to handle initialization.
It's a small change but it allows a subclass of Slop to initialize the
options (or some of the options). Example:

    class Opts < Slop
      def initialize config={}
        config[:help] = true

        super config do
          on 'v', 'verbose', 'Enable verbose mode'
        end
      end
    end

Now in my executable I do:

    opts = Opts.parse do
      ....
    end

Now my options are defined by a class. I could even make my class take a block
and futher allow defination of the options.