Support for required options
This commit introduces support for required options,
which are options that cause the parser to raise
a `MissingRequiredOption` exception if not present.

Options can be marked as required by passing
`required: true` in their configuration, and any errors
caused by missing required options can be suppressed via
`suppress_errors: true`.