Fix bug with separators around `help: false` opts
The separators rely on the indexes of the options before/after them, but
we ignore `help: false` options when building the help string, so all
of the positions are messed up.

This change avoids ignoring those options during the iteration, and
instead just doesn't append the option help to the string.

Fixes #253